* restructured some classes and templates to make them more uniform
