+++ public/modules/activity/poll/Activity_Poll.php
