+++ public/modules/activity/shout/Activity_Shout.php
