+++ public/modules/gw2api/widgets/gw2api_Widgets.php
