+++ public/themes/boilerplate/views/activity/event/tags/activity_event_tags_view.php
@@ -1,0 +1,1 @@
+{##signups##}
