+++ public/themes/boilerplate/views/activity/event/widgets/activity_event_scheduled_events.php
